# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: metarang
    description: REST API for metarang microservices. Legacy-compatible JSON responses served via Kong → service HTTP APIs.
    title: metarang API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: support-service HTTP direct (dev)
      url: http://localhost:8070
...
paths:
    /api/account/security:
        post:
            operationId: post_account_security
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Request account security verification
            tags:
                - Account Security
    /api/account/security/verify:
        post:
            operationId: post_account_security_verify
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Verify account security code
            tags:
                - Account Security
    /api/auth/callback:
        get:
            operationId: get_auth_callback
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: OAuth callback handler
            tags:
                - Auth
    /api/auth/logout:
        post:
            operationId: post_auth_logout
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Log out current user
            tags:
                - Auth
    /api/auth/me:
        get:
            operationId: get_auth_me
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get authenticated user profile
            tags:
                - Auth
    /api/auth/redirect:
        get:
            operationId: get_auth_redirect
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: Redirect to OAuth provider
            tags:
                - Auth
    /api/auth/register:
        post:
            operationId: post_auth_register
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: Start OAuth registration flow
            tags:
                - Auth
    /api/auth/validate:
        post:
            operationId: post_auth_validate
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: Validate bearer token
            tags:
                - Auth
    /api/bank-accounts:
        get:
            operationId: get_bank-accounts
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create bank accounts
            tags:
                - KYC
        post:
            operationId: post_bank-accounts
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create bank accounts
            tags:
                - KYC
...
    /api/events:
        get:
            operationId: get_events
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List user events
            tags:
                - Events
...
    /api/general-settings:
        get:
            operationId: get_general-settings
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get general settings
            tags:
                - Settings
...
    /api/kyc:
        get:
            operationId: get_kyc
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update KYC information
            tags:
                - KYC
        patch:
            operationId: patch_kyc
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update KYC information
            tags:
                - KYC
        put:
            operationId: put_kyc
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update KYC information
            tags:
                - KYC
...
    /api/mobile/send:
        post:
            operationId: post_mobile_send
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Send verification code to change mobile number
            tags:
                - Mobile
    /api/mobile/verify:
        post:
            operationId: post_mobile_verify
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Verify code and change mobile number
            tags:
                - Mobile
...
    /api/personal-info:
        get:
            operationId: get_personal-info
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update personal information
            tags:
                - Personal Info
        patch:
            operationId: patch_personal-info
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update personal information
            tags:
                - Personal Info
        put:
            operationId: put_personal-info
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update personal information
            tags:
                - Personal Info
    /api/privacy:
        get:
            operationId: get_privacy
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update privacy settings
            tags:
                - Settings
        post:
            operationId: post_privacy
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get or update privacy settings
            tags:
                - Settings
    /api/profile-limitations:
        post:
            operationId: post_profile-limitations
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Create profile limitation
            tags:
                - Profile
...
    /api/profile-photos:
        get:
            operationId: get_profile-photos
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or upload profile photos
            tags:
                - Profile
        post:
            operationId: post_profile-photos
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or upload profile photos
            tags:
                - Profile
...
    /api/search/isic-codes:
        get:
            operationId: get_search_isic-codes
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: Search ISIC codes
            tags:
                - Search
...
components:
    schemas:
...
        Error:
            properties:
                message:
                    type: string
            type: object
...
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http